Split water_sources.rs (525 lines) into water_sources/ subdirectory
Split crates/simulation/src/water_sources.rs into focused modules:
- types.rs: WaterSourceType enum, WaterSource component, constants
- systems.rs: update_water_sources, aggregate_water_source_supply, replenish_reservoirs
- plugin.rs: WaterSourcesPlugin registration
- tests.rs: all unit tests
- mod.rs: thin facade with re-exports
